Fix issue with the wrong offset by saving it to the db
The idea is that we will store the used offset in the request to Yelp's API.
This way we don't need to fetch the restaurants which were already discarded or
kept by the user.
The stored offset depends on the search preferences, if they change we will use
offset of 0.

Time spent around 0.5h.